Reorganize J3D structs (no change in weak order)
