+++ src/JSystem/J3DGraphBase/J3DMatBlock.cpp
@@ -2019,13 +2019,3 @@ void J3DTexGenBlockPatched::calcPostTexMtxWithoutViewMtx(f32 const (*param_0)[4]
         }
     }
 }
-
-void J3DTevBlock::diffTevReg() {}
-
-void J3DTevBlock::diffTevStageIndirect() {}
-
-void J3DTevBlock::diffTevStage() {}
-
-void J3DTevBlock::diffTexCoordScale() {}
-
-void J3DTevBlock::diffTexNo() {}
